rename to token_manager
