--- meetup/token_cache/__init__.py
@@ -1,6 +1,0 @@
-"""
-Easily obtain and cache OAuth 2.0 token from the Meetup API. Obtained tokens are
-stored both in memory and in Redis.
-"""
-
-from .token_cache import TokenCache
